requirements/base.txt
# Core framework
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
djoser==2.2.3

# Async / WebSocket
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Task queue
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1

# API utilities
django-filter==24.2
drf-spectacular==0.27.2

# Utilities
python-decouple==3.8
Pillow==10.4.0
django-storages==1.14.4